+++ src/components/pages/BlogsPage/index.tsx
@@ -7,8 +7,8 @@ import { connect, useDispatch, useSelector } from 'react-redux';
 import { Link, RouteComponentProps } from 'react-router-dom';
 import { Dispatch } from 'redux';
 import { ThunkDispatch } from 'redux-thunk';
-import { AppState } from '../../../redux/states/app-state';
-import { BlogState, fetchBlogs } from '../../../redux/states/blog-state';
+import { AppState } from '../../../redux/slices/app-state';
+import { BlogState, fetchBlogs } from '../../../redux/slices/blog-state';
 import { PrimaryAnkerButton } from '../../atoms/Button/index';
 import ScrollView from '../../atoms/ScrollView/index';
 import Wrapper from '../../atoms/Wrapper/index';
