+++ src/components/pages/IndexPage/index.tsx
@@ -4,8 +4,8 @@ import React, { useEffect } from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 import { Redirect, RouteComponentProps } from 'react-router-dom';
-import { AppState } from '../../../redux/states/app-state';
-import { UserState, fetchUser } from '../../../redux/states/user-state';
+import { AppState } from '../../../redux/slices/app-state';
+import { UserState, fetchUser } from '../../../redux/slices/user-state';
 import LoadingView from '../../molecules/LoadingView/index';
 import PageLayout from '../../templates/PageLayout/index';
 import WelcomePage from '../WelcomePage/index';
