+++ src/components/pages/SettingsPage/index.tsx
@@ -9,9 +9,9 @@ import { connect, useDispatch, useSelector } from 'react-redux';
 import { RouteComponentProps } from 'react-router';
 import { Link } from 'react-router-dom';
 import { ThunkDispatch } from 'redux-thunk';
-import { signOut } from '../../../redux/states/user-state';
-import { AppState } from '../../../redux/states/app-state';
-import { BlogState, fetchBlogs } from '../../../redux/states/blog-state';
+import { signOut } from '../../../redux/slices/user-state';
+import { AppState } from '../../../redux/slices/app-state';
+import { BlogState, fetchBlogs } from '../../../redux/slices/blog-state';
 import { Button } from '../../atoms/Button/index';
 import ScrollView from '../../atoms/ScrollView/index';
 import Wrapper from '../../atoms/Wrapper/index';
